cleanup the serialization code
